separate structure for boot node meta data
